biztrackr/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.14
jinja2==3.1.4
python-multipart==0.0.9
pydantic-settings==2.4.0
